\name{Cross.Parker.Consulting.node.region}
\docType{data}
\alias{Cross.Parker.Consulting.node.region}
\title{ Intra-organisational networks }
\description{ This dataset contains two intra-organizational networks from a consulting company (46 employees). These networks was used by Cross and Parker (2004). \cr\cr
In the first network, the ties are differentiated on a scale from 0 to 5 in terms of frequency of information or advice requests ("Please indicate how often you have turned to this person for information or advice on work-related topics in the past three months"). 0: I Do Not Know This Person; 1: Never; 2: Seldom; 3: Sometimes; 4: Often; and 5:Very Often. \cr\cr
In the second network, ties are differentiated in terms of the value placed on the information or advice received ("For each person in the list below, please show how strongly you agree or disagree with the following statement: In general, this person has expertise in areas that are important in the kind of work I do."). The weights in this network is also based on a scale from 0 to 5. 0: I Do Not Know This Person; 1: Strongly Disagree; 2: Disagree; 3: Neutral; 4: Agree; and 5: Strongly Agree.\cr\cr
In addition to the relational data, the dataset also contains information about the people (nodal attributes). The following attributes are known: the organisational level (1 Research Assistant; 2: Junior Consultant; 3: Senior Consultant; 4: Managing Consultant; 5: Partner), gender (1: male; 2: female), region (1: Europe; 2: USA), and location (1: Boston; 2: London; 3: Paris; 4: Rome; 5: Madrid; 6: Oslo; 7: Copenhagen).\cr\cr
See http://toreopsahl.com/datasets/ }
\usage{Cross.Parker.Consulting.node.region}
\format{ A vector with nodal attribute. }
\references{ Cross, R., Parker, A., 2004. The Hidden Power of Social Networks. Harvard Business School Press, Boston, MA.\cr
http://toreopsahl.com/datasets/ }
\keyword{datasets}