\name{Davis.Southern.women}
\docType{data}
\alias{Davis.Southern.women.2mode}
\alias{Davis.Southern.women.1mode.Cooccurance}
\alias{Davis.Southern.women.1mode.Newman}
\title{ Davis' Southern Women network }
\description{
This dataset was collected by Davis and colleague in the 1930s. It contains the observed attendance at 14 social events by 18 Southern women. This has been projected onto a co-occurance one-mode network, and a one-mode network based on Newman's (2001) method. }
\usage{
Davis.Southern.women.2mode
Davis.Southern.women.1mode.Cooccurance
Davis.Southern.women.1mode.Newman
}
\format{ A data frame with three columns. The first is the id of the sender; the second is the id of the receiver; and the third is the weight of the tie. }
\references{ Davis, A., Gardner, B. B., Gardner, M. R., 1941. Deep South. University of Chicago Press, Chicago, IL.\cr
http://toreopsahl.com/datasets/}
\author{ Tore Opsahl; http://toreopsahl.com }
\keyword{datasets}