\name{Freeman_EIES_3.Sym_sum}
\docType{data}
\alias{Freeman_EIES_3.Sym_sum}
\title{Symmetrised version of Freeman's EIES network (no. 3)}
\description{
EIES data sets 

The Electronic Information Exchange System (EIES) data were collected by Freeman and Freeman (1980). This data set is discussed also in Wasserman and Faust (1994). This is a network of 32 researchers who participated in an early study on the effects of electronic information exchange, a precursor of email communication. 

Coding for the network: 
The edges are weighted by the number of messages sent between scientists.
}
\usage{Freeman_EIES_3.Sym_sum}
\format{A data frame with three columns. The first is the sender, the second is the receiver, and the third is the number of messages.}
\source{Wasserman and Faust 1994, page 747}
\references{
Freeman, L.C., and Freeman, S.C., A semi-visible college: Structural effects on a social networks group. Pp. 77-85 in Henderson, M.M., and McNaughton, M.J. (eds.) Electronic Communication: Technology and Impacts Boulder, CO: Westview Press, 1980; 

Wasserman, Stanley, and Katherine Faust. Social Network Analysis: Methods and Applications. New York and Cambridge: Cambridge University Press, 1994. 
}
\keyword{datasets}