\name{OnlineComm.f1000.lnet}
\docType{data}
\alias{OnlineComm.f1000.lnet}
\title{Online community - extract }
\description{
The first 1000 messages between users in an online community.
}
\usage{OnlineComm.f1000}
\format{A data frame with four columns. The first is the time stamp at which the message was send; the second is the id of the sender; the third is the id of the receiver; and the fourth is the weight (always 1).}
\references{ Opsahl, T., Panzarasa, P., 2009. Clustering in weighted networks. Social Networks 31 (2), 155-163, doi: 10.1016/j.socnet.2009.02.002 \cr
http://toreopsahl.com/2009/04/03/article-clustering-in-weighted-networks/ }
\keyword{datasets}