\name{OnlineSocialNetwork.n1899}
\docType{data}
\alias{
OnlineSocialNetwork.n1899.net
OnlineSocialNetwork.n1899.lnet
}
\title{ Facebook-like Onlince Social Network }
\description{
This network is the Online Social Network-dataset used in my Ph.D. thesis. This network has also been described in Patterns and Dynamics of Users' Behaviour and Interaction: Network Analysis of an Online Community and used in Prominence and control: The weighted rich-club effect and Clustering in weighted networks. The network originate from an online communication network among students at University of California, Irvine. The edgelist includes the users that sent or received at least one message during that period (1,899). A total number of 59,835 online messages were sent among these over 20,296 directed ties.}
\usage{
OnlineSocialNetwork.n1899.net
OnlineSocialNetwork.n1899.lnet
}
\format{ A data frame with three columns. The first is the id of the sender; the second is the id of the receiver; and the third is the weight of the tie. }
\references{ Opsahl, T., Panzarasa, P., 2009. Clustering in weighted networks. Social Networks 31 (2), 155-163, doi: 10.1016/j.socnet.2009.02.002 \cr
http://toreopsahl.com/datasets/}
\author{ Tore Opsahl; http://toreopsahl.com }
\keyword{datasets}