\name{sample.network.1}
\docType{data}
\alias{sample.network.1}
\title{Sample network 1 in Opsahl and Panzarasa, Clustering in weighted networks. }
\description{
6 nodes with 6 edges, each carrying a weight between 1 and 4.
}
\usage{sample.network.1}
\format{A data frame with three columns. The first is the sender, the second is the receiver, and the third is the number of messages.}
\references{Opsahl. T., Panzarasa, P., 2009. Clustering in weighted networks. Social Networks. }
\author{ Tore Opsahl t.opsahl@qmul.ac.uk }
\keyword{datasets}