\name{OnlineSocialNetwork.n1899}
\docType{data}
\alias{
OnlineSocialNetwork.n1899.net
OnlineSocialNetwork.n1899.lnet
}
\title{ Facebook-like Online Social Network }
\description{
This network is the Facebook-like Social Network-dataset used in my Ph.D. thesis. This network has also been described in Patterns and Dynamics of Users' Behaviour and Interaction: Network Analysis of an Online Community and used in Prominence and control: The weighted rich-club effect and Clustering in weighted networks. The network originates from a virtual community among students at University of California, Irvine. The edgelists include the users that sent or received at least one message during that period (1,899). A total number of 59,835 online messages were sent among over 20,296 directed ties.}
\usage{
OnlineSocialNetwork.n1899.net
OnlineSocialNetwork.n1899.lnet
}
\format{ OnlineSocialNetwork.n1899.net: A data frame with three columns. The first column is the id of the sender, the second column is the id of the receiver, and the third column is the weight of the tie.\cr
OnlineSocialNetwork.n1899.lnet: A data frame with four columns. The first column is the timestamp, the second column is the id of the sender, the third column is the id of the receiver, and the fourth column is the weight of the tie (always 1).}
\references{ Opsahl, T., Panzarasa, P., 2009. Clustering in weighted networks. Social Networks 31 (2), 155-163, doi: 10.1016/j.socnet.2009.02.002 \cr
http://toreopsahl.com/datasets/}
\author{ Tore Opsahl; http://toreopsahl.com }
\keyword{datasets}